large model

+++ blib2.py
@@ -10,7 +10,7 @@ image = Image.open(requests.get(url, stream=True).raw).convert('RGB')
 processor = BlipProcessor.from_pretrained("Salesforce/blip-image-captioning-large")
 # by default `from_pretrained` loads the weights in float32
 # we load in float16 instead to save memory
-model = Blip2ForConditionalGeneration.from_pretrained("Salesforce/blip-image-captioning-large", torch_dtype=torch.float16) 
+model = Blip2ForConditionalGeneration.from_pretrained("Salesforce/blip-image-captioning-large") 
 
 device = "cuda" if torch.cuda.is_available() else "cpu"
 model.to(device)
